# feat(sendHtml/legacy): only check user agent for ie