--- a/package.json
+++ b/package.json
@@ -81,7 +81,6 @@
     "@americanexpress/vitruvius": "^2.0.0",
     "babel-preset-amex": "^3.2.0",
     "body-parser": "^1.19.0",
-    "browserslist-useragent": "^3.0.2",
     "chalk": "^3.0.0",
     "compression": "^1.7.4",
     "cookie-parser": "^1.4.4",