<!--- Provide a general summary of your changes in the Title above -->

## Description
<!--- Describe your changes in detail -->
Removed `browserslist-useragent` and implemented simple check for IE user agent.

## Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here. -->
This has both performance and security improvements. 
`browserslist-useragent` uses `useragent` to parse the user agent which in turn uses a huge list of [regular expressions](https://github.com/3rd-Eden/useragent/blob/master/lib/regexps.js) to parse the user agent string. This is both costly in performance and could potentially be introducing [security vulnerabilities](https://medium.com/@nitinpatel_20236/what-are-evil-regexes-7b21058c747e). 

## How Has This Been Tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, tests ran to see how -->
<!--- your change affects other areas of the code, etc. -->
In addition to unit and integration tests the One App server was started with `frank-llyod-root. autocannon was used to hit `localhost:3000/success` with 50 connections for a duration of 30 seconds.  

`autocannon localhost:3000/success -c 50 -d 30 --headers "Accept-Language= en-USen;q=0.5;" --headers "User-Agent= curl/7.5.0"` 

This was run three times before and after the changes. The fastest results are posted below

Before:

```
┌─────────┬───────┬───────┬───────┬────────┬──────────┬──────────┬───────────┐
│ Stat    │ 2.5%  │ 50%   │ 97.5% │ 99%    │ Avg      │ Stdev    │ Max       │
├─────────┼───────┼───────┼───────┼────────┼──────────┼──────────┼───────────┤
│ Latency │ 60 ms │ 64 ms │ 85 ms │ 126 ms │ 66.29 ms │ 10.08 ms │ 181.62 ms │
└─────────┴───────┴───────┴───────┴────────┴──────────┴──────────┴───────────┘
┌───────────┬─────────┬─────────┬─────────┬────────┬─────────┬────────┬─────────┐
│ Stat      │ 1%      │ 2.5%    │ 50%     │ 97.5%  │ Avg     │ Stdev  │ Min     │
├───────────┼─────────┼─────────┼─────────┼────────┼─────────┼────────┼─────────┤
│ Req/Sec   │ 651     │ 651     │ 759     │ 787    │ 748     │ 38.47  │ 651     │
├───────────┼─────────┼─────────┼─────────┼────────┼─────────┼────────┼─────────┤
│ Bytes/Sec │ 3.64 MB │ 3.64 MB │ 4.24 MB │ 4.4 MB │ 4.18 MB │ 215 kB │ 3.63 MB │
└───────────┴─────────┴─────────┴─────────┴────────┴─────────┴────────┴─────────┘

Req/Bytes counts sampled once per second.

22k requests in 30.06s, 125 MB read
```

After:

```
┌─────────┬───────┬───────┬───────┬────────┬─────────┬─────────┬───────────┐
│ Stat    │ 2.5%  │ 50%   │ 97.5% │ 99%    │ Avg     │ Stdev   │ Max       │
├─────────┼───────┼───────┼───────┼────────┼─────────┼─────────┼───────────┤
│ Latency │ 49 ms │ 51 ms │ 64 ms │ 102 ms │ 52.9 ms │ 7.47 ms │ 169.98 ms │
└─────────┴───────┴───────┴───────┴────────┴─────────┴─────────┴───────────┘
┌───────────┬─────────┬─────────┬─────────┬────────┬─────────┬────────┬─────────┐
│ Stat      │ 1%      │ 2.5%    │ 50%     │ 97.5%  │ Avg     │ Stdev  │ Min     │
├───────────┼─────────┼─────────┼─────────┼────────┼─────────┼────────┼─────────┤
│ Req/Sec   │ 870     │ 870     │ 940     │ 964    │ 935.64  │ 21.66  │ 870     │
├───────────┼─────────┼─────────┼─────────┼────────┼─────────┼────────┼─────────┤
│ Bytes/Sec │ 4.79 MB │ 4.79 MB │ 5.17 MB │ 5.3 MB │ 5.15 MB │ 119 kB │ 4.79 MB │
└───────────┴─────────┴─────────┴─────────┴────────┴─────────┴────────┴─────────┘

Req/Bytes counts sampled once per second.

28k requests in 30.06s, 154 MB read
```
## Types of Changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [ ] Documentation (adding or updating documentation)

## Checklist:
<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [ ] My change requires a change to the documentation and I have updated the documentation accordingly.
- [ ] These changes should be applied to a maintenance branch.
- [ ] This change requires cross browser checks.
- [x] Performance tests should be ran against the server prior to merging.
- [ ] This change impacts caching for client browsers.
- [ ] This change impacts HTTP headers.
- [ ] This change adds additional environment variable requirements for One App users.
- [ ] I have added the Apache 2.0 license header to any new files created.

## What is the Impact to Developers Using One App?
<!--- Please describe how your changes impacts developers using One App. -->
None