@mtomcal 2020-05-21
> This change requires cross browser checks.

I think we would want to use Browser Stack or something and check the multiple legacy IE's to make sure legacy bundles are served in the network tab to add confidence.

These results are exciting on the performance end!