@JAdshead 2020-05-21
> > This change requires cross browser checks.
> 
> I think we would want to use Browser Stack or something and check the multiple legacy IE's to make sure legacy bundles are served in the network tab to add confidence.
> 
> These results are exciting on the performance end!
I have cross referenced a few resources for all [valid IE user agents](http://www.useragentstring.com/pages/useragentstring.php?name=Internet+Explorer)
and tested making curl requests like the one below to also test the legacy bundles were being returned. 

`curl http://localhost:3000/success  -H "User-Agent: Mozilla/5.0 (compatible, MSIE 11, Windows NT 6.3; Trident/7.0; rv:11.0) like Gecko"`